requirements.txt
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.5.0
click==8.1.3
pytest==7.3.1
rich==13.3.5
pyyaml==6.0
tenacity==8.2.2
google-search-results==2.4.2
pytest-mock==3.10.0
pytest-cov==4.1.0 